Fixed compilation without "wasm-bindgen" feature
